Moved checkout tests from users spec to orders. Also added test to update inventory when user checks out. Also updates when Cancelled
